Еще продвинулся, повыносил в функции многое, осталось в других доделать
